+++ RELEASES.md
@@ -1,5 +1,8 @@
 ## 0.0.12-SNAPSHOT/20211007-14:03:32
 
+- Features:
+  - ebRigol math - setup math operator source1 source2 scale offset
+
 - Fixes:
   - ./ebUnit : ebUnit_main not found, core dump in no device found
   - ./ebRigol: has no attribute 'ebRigol_main'